Optimize Grover's Algorithm with new features (#53)
